Fix cumulativeUserMemory initial calculation & use double for cumulativeUserMemory consistently

+++ presto-main/src/main/java/com/facebook/presto/execution/QueryStats.java
@@ -269,7 +269,7 @@ public class QueryStats
         int blockedDrivers = 0;
         int completedDrivers = 0;
 
-        long cumulativeUserMemory = 0;
+        double cumulativeUserMemory = 0;
         long userMemoryReservation = 0;
         long totalMemoryReservation = 0;
